tablet-case-background-workflow/requirements-app.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
jinja2==3.1.6
python-multipart==0.0.21
psutil==7.1.2
Pillow==12.2.0
numpy==2.3.4
